gnome-continuous-journal-log-hack: New program
I want to log to journald from platform-independent test script
injected from gnome-continuous.  I don't want to write native testing
code since then I have to care about architecture which is annoying.

libgsystem of course has logging API but it's designed to be embedded
with your app, which is hard/annoying for injecting overlay binaries.

So...that leaves us with this hack, until we get structured logging in
GLib and unicorns for everyone.